pyproject.toml
[tool.poetry]
name = "sphinx-essearch"
version = "0.1.0"
description = ""
authors = ["yamachi"]
packages = [{include = "sphinx_essearch"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.26.42"
SQLAlchemy = "^1.4.42"
python-dotenv = "^0.21.0"
opensearch-py = "^2.0.1"
sphinxcontrib-websupport = "^1.2.4"
sphinx = "^6.0.0"
flask = "^2.2.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^0.991"
black = "^22.12.0"
pyproject-flake8 = "^6.0.0.post1"
boto3-stubs = {extras = ["essential"], version = "^1.26.42"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
ignore = "E203"
